Added getfeatureinfo and getattributes
